// sourcecode/BTrak.Packages/snova-payroll-widgets/module/module-components/src/lib/payrollwidgetsandreports/models/creditordetailsmodel.ts
export class CreditorDetailsModel{
    employeeCreditorDetailsId: string;
    branchId: string;
    accountNumber: string;
    accountName: string;
    ifScCode: string;
    bankName: string;
    branchName: string;
    isArchived: boolean;
    bankId: string;
    timeStamp: any;
    mobileNo: number;
    email: string;
    useForPerformaInvoice: boolean;
    panNumber: string;
}